+++ database/migrations/2025_05_06_232812_add_returned_product_and_product_problem_notes_applied_status.php
@@ -44,10 +44,14 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('returned_product_notes', function (Blueprint $table) {
-            $table->dropColumn('applied_status');
+            if (Schema::hasColumn('returned_product_notes', 'applied_status')) {
+                $table->dropColumn('applied_status');
+            }
         });
         Schema::table('product_problem_notes', function (Blueprint $table) {
-            $table->dropColumn('applied_status');
+            if (Schema::hasColumn('product_problem_notes', 'applied_status')) {
+                $table->dropColumn('applied_status');
+            }
         });
     }
 };
